New gesture, tweaked gestures and debug-flag for the camera
Added two versions of a new gesture (gust, with and without a charging
period). Tweaked hurricane to accomodate a max-force (doesn't feel good
yet though).  Added a public bool on the camera-script, so it can be
frozen for debug-purposes. Fixed a few float-typed-as-int-bugs causing
major havoc. Also wasted hours of my life unsuccesfully fixing a
null-pointer in the RelicController. Had coffee though, so not entirely
a waste. ...Is anyone even reading these things?